requirements.txt
apache-airflow 
# psycopg2 
mlflow[extras] 
scikit-learn 
pandas 
numpy 